+++ server/config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"errors"
 	"fmt"
+	"net/netip"
 	"path"
 	"time"
 
@@ -65,6 +66,9 @@ const (
 
 	// DefaultEVMMinTip is the default minimum priority fee for the mempool
 	DefaultEVMMinTip = 0
+
+	// DefaultGethMetricsAddress is the default port for the geth metrics server.
+	DefaultGethMetricsAddress = "127.0.0.1:8100"
 
 	// DefaultGasCap is the default cap on gas that can be used in eth_call/estimateGas
 	DefaultGasCap uint64 = 25_000_000
@@ -145,6 +149,8 @@ type EVMConfig struct {
 	EVMChainID uint64 `mapstructure:"evm-chain-id"`
 	// MinTip defines the minimum priority fee for the mempool
 	MinTip uint64 `mapstructure:"min-tip"`
+	// GethMetricsAddress is the address the geth metrics server will bind to. Default 127.0.0.1:8100
+	GethMetricsAddress string `mapstructure:"geth-metrics-address"`
 }
 
 // JSONRPCConfig defines configuration for the EVM RPC server.
@@ -213,6 +219,7 @@ func DefaultEVMConfig() *EVMConfig {
 		EVMChainID:              DefaultEVMChainID,
 		EnablePreimageRecording: DefaultEnablePreimageRecording,
 		MinTip:                  DefaultEVMMinTip,
+		GethMetricsAddress:      DefaultGethMetricsAddress,
 	}
 }
 
@@ -220,6 +227,10 @@ func DefaultEVMConfig() *EVMConfig {
 func (c EVMConfig) Validate() error {
 	if c.Tracer != "" && !strings.StringInSlice(c.Tracer, evmTracers) {
 		return fmt.Errorf("invalid tracer type %s, available types: %v", c.Tracer, evmTracers)
+	}
+
+	if _, err := netip.ParseAddrPort(c.GethMetricsAddress); err != nil {
+		return fmt.Errorf("invalid geth metrics address %q: %w", c.GethMetricsAddress, err)
 	}
 
 	return nil