+++ server/flags/flags.go
@@ -70,6 +70,7 @@ const (
 	EVMEnablePreimageRecording = "evm.cache-preimage"
 	EVMChainID                 = "evm.evm-chain-id"
 	EVMMinTip                  = "evm.min-tip"
+	EvmGethMetricsAddress      = "evm.geth-metrics-address"
 )
 
 // TLS flags